// frontend/packages/kommon/src/components/RoleForm/types.ts
export interface I18NPermissionTree {
  folders: string
  documents: string
  page_management: string
  categories: string
  shares: string
  users: string
  roles: string
  groups: string
  view: string
  create: string
  update: string
  move: string
  delete: string
  download: string
  upload: string
  extract: string
  rotate: string
  reorder: string
  select: string
  all_versions: string
  only_last_version: string
  title: string
  custom_fields: string
  tags: string
  category: string
}

export interface I18NCollapseButton {
  collapseAll: string
  expandAll: string
}

export interface I18NCheckButton {
  checkAll: string
  uncheckAll: string
}